change to float and add integration thread

+++ include/remy_robot_control/angle.h
@@ -5,7 +5,7 @@ namespace remy_robot_control {
 /* A custom class to handle angles properly (-PI, PI). One may choose the angle
  * Limits.
 */
-template <class T = double>
+template <class T = float>
 class Angle {
   T value_;
   T min_;
